BF: Two simple bug-fixes.
In model: the residual needs to be defined as what remains of the signal when
you subtract the model prediction.

In viz: Dealing with the cmap argument separately and removing it from the
kwarg, so that it doesn't get passed to plot3D (which doesn't know what to do
with it).