ENH: New module for parallel execution of model fitting.
This requires a cluster running sge. Once the wrapper is completed, you need to
log onto the cluster and run batch_sge.sh, which submits everything over qsub.

This could be simplified a bit, so that the call there is also made over
paramiko, but so far I haven't been able to set the environment variables
properly through the paramiko connection. Will try that when everyone gets back
from holidays.